Fix README and rename Projects section
