+++ my-portfolio/src/App.jsx
@@ -6,7 +6,7 @@ import { Navbar } from './components/Navbar.jsx';
 import { MobileMenu } from './components/MobileMenu.jsx';
 import { Home } from './components/sections/Home.jsx';
 import { About } from './components/sections/About.jsx';
-import { Projects } from './components/sections/Porjects.jsx';
+import { Projects } from './components/sections/Projects.jsx';
 import { Contact } from './components/sections/Contact.jsx';
 
 function App() {
